\name{drvisits}
          \docType{data}
          \alias{drvisits}
          \title{Count Response Model}
          \description{
	    In this example we use data (drvisits.dat) from the German
	    Socio-Economic Panel (SEOP) on the self reported number of
	    visits by women to the doctor, just
	    before a major health care reform in 1997 and just after the
	    reform. The health reform raised prescription charges by
	    200% and imposed limits on fees
	    charged by doctors. This data was analysed by Winkelmann
	    (2004), Rabe-Hesketh and Skrondal (2005, Ch 6) who note that
	    it is interesting to establish if the
	    health reforms tended to reduce visits by the women to the
	    doctor. Following Rabe-Hesketh and Skrondal (2005, Ch 5) we
	    consider the subset of women who were
	    working full time in the 1996 and 1998 panels of SEOP.

	    Rabe-Hesketh and Skrondal (2005, Ch 5) show that fewer than
	    50\% of the women provide data for both the 1996 and 1998
	    waves. For the purposes of this analysis and like
	    Rabe-Hesketh and Skrondal (2005, Ch 5) we treat the
	    missing data as ignorable and use all the observed responses.

          }
          \usage{drvisits}
          \format{A data frame containing 2227 rows and 10 columns.}
          \source{German Socio-Economic Panel (SEOP)}


	  \references{
	    Rabe-Hesketh, S., and Skrondal, A., (2005), \emph{Multilevel
	      and Longitudinal Modelling using Stata}, Stata Press,
	    Stata Corp, College Station, Texas.

	    Winkelmann, R., (2004), \emph{Healthcare reform and thye
	      number of doctor visits: an econometric analysis}, Journal
	    of Applied Econometrics, 19, 455-472.	    
          }

	  
          \keyword{datasets}
