\name{growth}
          \docType{data}
          \alias{growth}
          \title{Continuos Response Growth Model}
          \description{
	    Snijders & Bosker, (1999) analysed the development over time
	    of teacher evaluations.  Starting from the first year of
	    their career, teachers were evaluated on their interpersonal
	    behaviour in the classroom.  This happened
	    repeatedly, at intervals of about one year.  In this
	    example, results are presented
	    about the 'proximity' dimension, representing the degree of
	    cooperation or closeness between a teacher and his or her
	    students.  The higher the proximity score
	    of a teacher, the more cooperation is perceived by his or her students.
          }
          \usage{growth}
          \format{A data frame containing 153 rows and 8 columns.}
          \source{Snijders & Bosker, (1999)}

	  \references{
	    Snijders, T. A. B. and Bosker, R. J. (1999). \emph{Multilevel Analysis}. London: Sage.
          }

	  
          \keyword{datasets}
