\name{schiz}
          \docType{data}
          \alias{schiz}
          \title{Ordered Response Model}
          \description{
	    In this example we analyse ordinal data on changes in the
	    severity of Schizophrenia in data from the National
	    Institute of Mental Health (NIMH), see Rabe-Hesketh and
	    Skrondal (2005, Ch 5). Patients were randomly allocated to 1
	    of 4 treatments (Placebo, Chlorpromazine, Fluphenazine,
	    Thioridazine) and examined weekly for up to 6 weeks. The
	    data have previously been analysed by Hedeker
	    and Gibbons (1996a), Gibbons et al. (1998) and Gibbons and
	    Hedeker (1994). In this example we follow Rabe-Hesketh and
	    Skrondal (2005, Ch 5) and combine
	    the treatments Chlorpromazine, Fluphenazine and Thioridazine. 
          }
          \usage{schiz}
          \format{A data frame containing 1603 rows and 8 columns.}
          \source{National Institute of Mental Health (NIMH)}

	  \references{
	    Gibbons, R. D.,  Hedeker, C., Waterneaux, C., & Davis, J.M.,
	    (1988), \emph{Random regression models: A comprehensive approach
	    to the analysis of longitudinal psychiatric data}, Psychopharmacology Bulletin, 24, 438-443
	    
	    Gibbons, R. D., & Hedeker, C., (1994), \emph{Application of random effects probit regression models}, Journal of Consulting and Clinical Psyschology.
	    
	    Hedeker, D., & Gibbons, R.D., (1996), \emph{Applied longitudinal data anlaysis}, Chichester, UK, Wiley
	    
	    Rabe-Hesketh, S., and Skrondal, A., (2005), \emph{Multilevel and Longitudinal Modelling using Stata, Stata Press}, Stata Corp, College Station, Texas.
          }

	  
          \keyword{datasets}
