\name{tu}
          \docType{data}
          \alias{tu}
          \title{Binary Response Model}
          \description{
	    As part of the Social Change and Economic Life Initiative,
	    funded by the U.K. Economic and Social Research Council in
	    the mid-1980s, complete work history information was
	    collected on 1000 individuals then resident in Rochdale,
	    Lancashire. The work histories were collected in the form of
	    work events, with a start and finish date as well as much
	    detailed information on the employment state - the type and
	    conditions of employment for each event. A subset of data
	    consisting of all work events for males was then extracted,
	    excluding any work event with missing data or relating to a
	    period of unemployment. This produced 1633 work events on
	    307 cases. Of interest was the relationship of trade-union
	    membership to social class (6 levels), number of employees
	    (5 levels), supervisory
	    responsibility (3 levels), age of individual at start of work event,
	    hours worked (3 levels), work with mainly one sex (4 levels)
	    and promotion prospects (2 levels).
          }
          \usage{tu}
          \format{A data frame containing 1633 rows and 11 columns.}
          \source{Social Change and Economic Life Initiative}


	  \references{
	    Barry, J. T., Francis, B. J., and Davies, R. B., (1989),
	    \emph{SABRE: Software for the Analysis of Binary Recurrent
	      Events},
	    In Lecture Notes in Statistical Modelling 57,
	    Decarli, A., Francis, B. J., Gilchrist, R. and Seeber
	    G. U. H. (editors).
	    Proceedings of Statistical Modelling, Trento.
	    Springer-Verlag
          }

	  
          \keyword{datasets}
